% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy_function.R
\name{accuracy}
\alias{accuracy}
\title{BMI585 final project: including accuracy}
\usage{
accuracy(pred, true)
}
\arguments{
\item{pred:}{the prediction array}

\item{true:}{the true array}
}
\value{
sen: accuracy
}
\description{
This functions accept two binary numerical arrays
}
\examples{
pred <- c(1,1,1)
true <- c(1,1,0)
accuracy(pred,true)
}
