% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjusted_R2_function.R
\name{adjusted_R2}
\alias{adjusted_R2}
\title{BMI585 final project: adjusted R^2 measure for the regression model which has penalty on the number of parameters}
\usage{
adjusted_R2(y, y_d, n)
}
\arguments{
\item{y:}{the numerical array for the ground truth}

\item{y_d:}{the numerical array for the regression}
}
\value{
R_s: the adjusted R^2 measure
}
\description{
This functions accept two numerical arrays
}
\examples{
y <- c(1,1,1)
y_d <- c(1,1,0.5)
adjusted_R2(y,y_d)
}
