% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhAdjust_function.R
\name{bhAdjust}
\alias{bhAdjust}
\title{BMI585 final project: the Bonferroni-Holm function for multiple comparison adjustment}
\usage{
bhAdjust(p, alpha)
}
\arguments{
\item{p:}{the numerical array for multiple comparison}

\item{alpha:}{number of values in the first array}
}
\value{
p: post-hoc power
}
\description{
This functions accept one numerical array and a numerical value
}
\examples{
p<- c(0.0025, 0.0050, 0.0075, 0.0100, 0.0125, 0.0150, 0.0175, 0.0200, 0.0225, 0.0250)
alpha <- 0.05
bhAdjust(p,alpha)
}
