% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOXMuller_function.R
\name{BOXMuller}
\alias{BOXMuller}
\title{BMI585 final project: BOXMuller creates n samples from the Box-Muller transformation}
\usage{
BOXMuller(n)
}
\arguments{
\item{n:}{n samples}
}
\value{
data.frame(x_1,x_2): x_1 and x_2 are samples from the Box-Muller transformation
}
\description{
This functions accept one numerical value
}
\examples{
BOXMuller(500)
}
