% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffectSize_function.R
\name{EffectSize}
\alias{EffectSize}
\title{BMI585 final project: EffectSize calculate the cohen's effective size}
\usage{
EffectSize(t, g)
}
\arguments{
\item{t:}{treatment array}

\item{g:}{control array}
}
\value{
size: effctive size
}
\description{
This functions accept two numerical array
}
\examples{
t <- c(1,1,1)
g <- c(0,0,0)
EffectSize(t,g)
}
