% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1_function.R
\name{f1}
\alias{f1}
\title{BMI585 final project: including f1}
\usage{
f1(pred, true)
}
\arguments{
\item{pred:}{the prediction array}

\item{true:}{the true array}
}
\value{
sen: f1
}
\description{
This functions accept two binary numerical arrays
}
\examples{
pred <- c(1,1,1)
true <- c(1,1,0)
f1(pred,true)
}
