% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_sample_size_function.R
\name{min_sample_size}
\alias{min_sample_size}
\title{BMI585 final project: find the minimum sample size}
\usage{
min_sample_size(power)
}
\arguments{
\item{power:}{power for statistics}
}
\value{
result: minimum sample size
}
\description{
This functions accept a numerical value
}
\examples{
power <- 1
min_sample_size(power)
}
