% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcApprox_function.R
\name{pcApprox}
\alias{pcApprox}
\title{BMI585 final project: pcApprox function dowmscale the data's dimensionlity}
\usage{
pcApprox(t, npc)
}
\arguments{
\item{t:}{a numerical dataframe for dimensionality reduction}
}
\value{
the_result[,1:npc]: a dataframe which is the PCA conversion of the input x with npc dimensions
}
\description{
This functions accept a numberical dataframe
}
\examples{
data(iris)
t<- iris[1:100,1:4]
pred <- pcApprox(t,2)
}
