% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcLollipop_function.R
\name{pcLollipop}
\alias{pcLollipop}
\title{BMI585 final project: pcLollipop function show a plot of PCA components with their importance}
\usage{
pcLollipop(t)
}
\arguments{
\item{t:}{a numerical dataframe for PCA analysis}
}
\value{
theplot: The plot is directly shown up
}
\description{
This functions accept a numberical dataframe
}
\examples{
data(iris)
t<- iris[1:100,1:4]
theplot<-pcLollipop(t)
theplot
}
