% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postHocPower_function.R
\name{postHocPower}
\alias{postHocPower}
\title{BMI585 final project: the posthoc power for comparison of two array}
\usage{
postHocPower(d, n1, n2)
}
\arguments{
\item{d:}{mean difference}

\item{n1:}{number of values in the first array}

\item{n2:}{number of values in the second array}
}
\value{
p: post-hoc power
}
\description{
This functions accept 3 numerical values
}
\examples{
d <- 1
n1 <- 20
n2 <- 20
postHocPower(d,n1,n2)
}
