% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2_function.R
\name{R2}
\alias{R2}
\title{BMI585 final project: R^2 measure for the regression model}
\usage{
R2(y, y_d)
}
\arguments{
\item{y:}{the numerical array for the ground truth}

\item{y_d:}{the numerical array for the regression}
}
\value{
R_s: the R^2 measure
}
\description{
This functions accept two numerical arrays
}
\examples{
y <- c(1,1,1)
y_d <- c(1,1,0.5)
R2(y,y_d)
}
