% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_function.R
\name{sensitivity}
\alias{sensitivity}
\title{BMI585 final project: including sensitivity}
\usage{
sensitivity(pred, true)
}
\arguments{
\item{pred:}{the prediction array}

\item{true:}{the true array}
}
\value{
sen: sensitivity
}
\description{
This functions accept two binary numerical arrays
}
\examples{
pred <- c(1,1,1)
true <- c(1,1,0)
sensitivity(pred,true)
}
