% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificity_function.R
\name{specificity}
\alias{specificity}
\title{BMI585 final project: including specificity}
\usage{
specificity(pred, true)
}
\arguments{
\item{pred:}{the prediction array}

\item{true:}{the true array}
}
\value{
sen: specificity
}
\description{
This functions accept two binary numerical arrays
}
\examples{
pred <- c(1,1,1)
true <- c(1,1,0)
specificity(pred,true)
}
