% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_sided_t_function.R
\name{two_sided_t}
\alias{two_sided_t}
\title{BMI585 final project: two_sided_t convert a test statistic into the area under the t-density}
\usage{
two_sided_t(t, df)
}
\arguments{
\item{t:}{the upper bound |t|}

\item{df:}{a numerical array}
}
\value{
result: an array of area < |t|
}
\description{
This functions accept a numerical value and and dataframe
}
\examples{
df <- c(1,1,1)
c <- 0.5
two_sided_t(t,df)
}
