% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_sided_z_function.R
\name{two_sided_z}
\alias{two_sided_z}
\title{BMI585 final project: two_sided_z convert a test statistic into the area under the z-density}
\usage{
two_sided_z(z)
}
\arguments{
\item{t:}{the upper bound |z|}
}
\value{
result: an array of area < |z|
}
\description{
This functions accept a numerical value and and dataframe
}
\examples{
z <- 0.5
two_sided_z(z)
}
