% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unscale_function.R
\name{Unscale}
\alias{Unscale}
\title{BMI585 final project: unscale function unscales the import data}
\usage{
Unscale(scale_ob)
}
\arguments{
\item{scale_ob:}{a scale object}
}
\value{
result: the class with center, scale and the original unscale data
}
\description{
This functions accept a class from the function scale
}
\examples{
x <- matrix(1:10, ncol = 2)
scale_ob <- scale(x)
Unscale(scale_ob)
}
