% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_sample_size_function.R
\name{min_sample_size}
\alias{min_sample_size}
\title{BMI585 final project: find the minimum sample size}
\usage{
min_sample_size(effect_size)
}
\arguments{
\item{effect_size:}{difference in means between the two groups divided by the sd}
}
\value{
result: minimum sample size
}
\description{
This functions accept a numerical value
}
\examples{
effect_size <- 1
min_sample_size(effect_size)
}
